+++ demos/chinook/src/main/java/module-info.java
@@ -46,6 +46,8 @@ module is.codion.framework.demos.chinook {
 					with is.codion.demos.chinook.domain.ChinookImpl;
 	provides is.codion.common.rmi.server.Authenticator
 					with is.codion.demos.chinook.server.ChinookAuthenticator;
+	provides is.codion.common.db.database.SessionContext
+					with is.codion.demos.chinook.server.ChinookSessionContext;
 	provides is.codion.common.utilities.resource.Resources
 					with is.codion.demos.chinook.i18n.ChinookResources;
 	// tag::entityObjectMapper[]